Feat: 수정 및 삭제의 사용자 확인 기능

+++ src/main/kotlin/com/noreabang/strawberryrabbit/domain/comment/service/CommentService.kt
@@ -8,6 +8,7 @@ import com.noreabang.strawberryrabbit.domain.feed.repository.FeedRepository
 import com.noreabang.strawberryrabbit.domain.member.repository.MemberRepository
 import com.noreabang.strawberryrabbit.infra.exception.ModelNotFoundException
 import org.springframework.data.repository.findByIdOrNull
+import org.springframework.security.access.AccessDeniedException
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -27,14 +28,25 @@ class CommentService(
         ).toResponse()
     }
 
-    fun updateComment(commentId: Long, request: CommentRequest): CommentResponse {
+    fun updateComment(commentId: Long, memberId: Long?, request: CommentRequest): CommentResponse {
+        val member = memberId.let { memberRepository.findByIdOrNull(it) }
         val comment = commentRepository.findByIdOrNull(commentId) ?: throw ModelNotFoundException("Comment", commentId)
-            comment.content = request.content
+
+        if (comment.member != member) {
+            throw AccessDeniedException("Not access")
+        }
+
+        comment.content = request.content
             return commentRepository.save(comment).toResponse()
     }
 
-    fun deleteComment(commentId: Long) {
+    fun deleteComment(commentId: Long, memberId: Long?) {
+        val member = memberId.let { memberRepository.findByIdOrNull(it) }
         val comment = commentRepository.findByIdOrNull(commentId) ?: throw ModelNotFoundException("Comment", commentId)
+
+        if (comment.member != member) {
+            throw AccessDeniedException("Not access")
+        }
         commentRepository.delete(comment)
     }
 }
